[Assignment 3] Python implementation
Implemented Karger's algorithm to find the mincut of an undirected
graph. The algorithm does not follow the Karger's algorithm exactly in
the clause of running it polynimial time. This implementation runs it
for a linear time (equal to number of vertices) but gives the correct
output because linear times running of the algorithm also reduces the
probability of failure to quite small.