[Princeton Algorithms - I] Implementing Queues
Got perfect score (100%) in the assignment with these modifications.
Modification does minor change in the implementation of sample() API.
Then used this sample() function for randomness in dequeue() API instead
for shuffle() API. This made the time complexity of dequeue of a single
element of constant order from linear order.

Signed-off-by: Siddharth Goel <siddharth98391@gmail.com>